pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gait-research-platform"
version = "0.1.0"
description = "AI-driven gait research platform MVP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "omegaconf>=2.3",
  "pandas>=2.2",
  "pyarrow>=15.0",
  "pyyaml>=6.0",
  "scikit-learn>=1.4",
  "scipy>=1.12",
  "seaborn>=0.13",
  "torch>=2.2",
]

[project.optional-dependencies]
agent = ["openai>=1.30"]
dev = ["pytest>=8.0"]
umap = ["umap-learn>=0.5.5"]
video = ["decord>=0.6.0", "opencv-python>=4.9"]

[tool.setuptools.packages.find]
include = ["gait_research_platform*"]
